refactor: move insturmentation in plugin
